+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+## v4.0.0 2021 July 28
+
+-   Prior `fetchExclusiveCompatibleESVersionsForNodeVersions` is now `fetchAllCompatibleESVersionsForNodeVersions`
+-   `fetchExclusiveCompatibleESVersionsForNodeVersions` now only returns the minimum amount of ECMAScript versions that are required to support the specified Node.js releases, which was its intent
+
 ## v3.0.0 2021 July 28
 
 -   Same API, however additional properties and methods for simplifying the identification of which ECMAScript versions are actually compatible